improve import project logics
